+++ Sources/VComponents/Components/State Pickers/Toggle Button (Rectangular)/VRectangularToggleButtonState.swift
@@ -11,13 +11,11 @@ import VCore
 // MARK: - V Rectangular Toggle Button State
 /// Enumeration that represents state, such as `off` or `on`.
 @available(tvOS, unavailable)
-@available(watchOS, unavailable)
 @available(visionOS, unavailable)
 public typealias VRectangularToggleButtonState = GenericState_OffOn
 
 // MARK: - V Rectangular Toggle Button Internal State
 /// Enumeration that represents state, such as `off`, `on`, `pressed`, or `disabled`.
 @available(tvOS, unavailable)
-@available(watchOS, unavailable)
 @available(visionOS, unavailable)
 public typealias VRectangularToggleButtonInternalState = GenericState_OffOnPressedDisabled