+++ Sources/VComponents/Components/State Pickers/Toggle Button (Wrapped)/VWrappedToggleButton.swift
@@ -46,7 +46,6 @@ import VCore
 ///     }
 ///
 @available(tvOS, unavailable) // Doesn't follow HIG
-@available(watchOS, unavailable) // ???
 @available(visionOS, unavailable) // Doesn't follow HIG
 public struct VWrappedToggleButton<Label>: View where Label: View {
     // MARK: Properties - UI Model
@@ -232,13 +231,14 @@ public struct VWrappedToggleButton<Label>: View where Label: View {
     private func playHapticEffect() {
 #if os(iOS)
         HapticManager.shared.playImpact(uiModel.haptic)
+#elseif os(watchOS)
+        HapticManager.shared.playImpact(uiModel.haptic)
 #endif
     }
 }
 
 // MARK: - Helpers
 @available(tvOS, unavailable)
-@available(watchOS, unavailable)
 @available(visionOS, unavailable)
 extension VWrappedToggleButtonInternalState {
     fileprivate var isPressedOffPressedOn: Bool {
@@ -255,7 +255,7 @@ extension VWrappedToggleButtonInternalState {
 // MARK: - Preview
 #if DEBUG
 
-#if !(os(tvOS) || os(watchOS) || os(visionOS))
+#if !(os(tvOS) || os(visionOS))
 
 #Preview("*", body: {
     struct ContentView: View {