+++ Sources/VComponents/Components/State Pickers/Toggle Button (Wrapped)/VWrappedToggleButtonState.swift
@@ -11,13 +11,11 @@ import VCore
 // MARK: - V Wrapped Toggle Button State
 /// Enumeration that represents state, such as `off` or `on`.
 @available(tvOS, unavailable)
-@available(watchOS, unavailable)
 @available(visionOS, unavailable)
 public typealias VWrappedToggleButtonState = GenericState_OffOn
 
 // MARK: - V Wrapped Toggle Button Internal State
 /// Enumeration that represents state, such as `off`, `on`, `pressed`, or `disabled`.
 @available(tvOS, unavailable)
-@available(watchOS, unavailable)
 @available(visionOS, unavailable)
 public typealias VWrappedToggleButtonInternalState = GenericState_OffOnPressedDisabled